+++ src/lib.rs
@@ -287,13 +287,15 @@ pub mod internals {
 #[doc(hidden)]
 #[cfg(feature = "_cargo_insta_internal")]
 pub mod _cargo_insta_support {
-    pub use crate::env::{
-        Error as ToolConfigError, OutputBehavior, SnapshotUpdate, TestRunner, ToolConfig,
-    };
     pub use crate::{
+        env::{
+            Error as ToolConfigError, OutputBehavior, SnapshotUpdate, TestRunner, ToolConfig,
+            UnreferencedSnapshots,
+        },
         output::{print_snapshot, print_snapshot_diff},
         snapshot::PendingInlineSnapshot,
         snapshot::SnapshotContents,
+        utils::is_ci,
     };
 }